[CLI] Make it simplier
